doc/radosgw: Improve language and formatting in config-ref.rst
The "Topic persistency settings" section includes convoluted
and incomplete sentences so attempt to improve the
language. The section lacks context so add a link to the
relevant documentation about Notifications.

The "Garbage Collection settings" note includes a sentence
that looks incomplete so try to improve the language.

Also improve formatting: indentation of a note and line
wrapping.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>