doc/radosgw: Small improvements in s3-notification-compatibility.rst
Attempt a small fix to a grammatical error in a sentence.
It should also refer to "below" and not "above", probably.

End full sentences in full stops.

Indent an unordered list consistently so that it renders consistently
with the same bullets.
Also indent all various blocks at the same columns consistently.

Wrap lines before column 80 while we're at it.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>